requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
mysql-connector-python
